+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+## v2.0.0
+
+ * Added support for version 2.0 of Mailchimp API
+ * Added HTTP POST method support to HTTP Adapters
+ * Dropped Socket HTTP Adapter
+ * Added Documentation for Mailchimp API and Mailchimp Export API
+
 ## v1.0.0
 
-* Mailchimp API connector
-* Mailchimp Export API connector
+ * Mailchimp API connector
+ * Mailchimp Export API connector